+++ app/catalog/service.go
@@ -3,12 +3,14 @@ package catalog
 import (
 	"log"
 
+	"github.com/mytheresa/go-hiring-challenge/app/api"
 	"github.com/mytheresa/go-hiring-challenge/models"
 )
 
 type Product struct {
-	Code  string  `json:"code"`
-	Price float64 `json:"price"`
+	Code     string                `json:"code"`
+	Price    float64               `json:"price"`
+	Category *api.CategoryResponse `json:"category"`
 }
 
 type IProductRepository interface {
@@ -44,9 +46,18 @@ func (s *CatalogService) ListProducts() ([]Product, error) {
 func toCatalogProducts(products []models.Product) []Product {
 	mapped := make([]Product, len(products))
 	for i, p := range products {
+		var category *api.CategoryResponse
+		if p.Category != nil {
+			category = &api.CategoryResponse{
+				Code: p.Category.Code,
+				Name: p.Category.Name,
+			}
+		}
+
 		mapped[i] = Product{
-			Code:  p.Code,
-			Price: p.Price.InexactFloat64(),
+			Code:     p.Code,
+			Price:    p.Price.InexactFloat64(),
+			Category: category,
 		}
 	}
 	return mapped