+++ app/categories/service.go
@@ -1,20 +1,16 @@
 package categories
 
 import (
+	"github.com/mytheresa/go-hiring-challenge/app/api"
 	"github.com/mytheresa/go-hiring-challenge/models"
 )
-
-type Category struct {
-	Code string `json:"code"`
-	Name string `json:"name"`
-}
 
 type ICategoryRepository interface {
 	GetAllCategories() ([]models.Category, error)
 }
 
 type ICategoriesService interface {
-	ListCategories() ([]Category, error)
+	ListCategories() ([]api.CategoryResponse, error)
 }
 
 type CategoriesService struct {
@@ -25,7 +21,7 @@ func NewCategoriesService(repo ICategoryRepository) *CategoriesService {
 	return &CategoriesService{repo: repo}
 }
 
-func (s *CategoriesService) ListCategories() ([]Category, error) {
+func (s *CategoriesService) ListCategories() ([]api.CategoryResponse, error) {
 	categories, err := s.repo.GetAllCategories()
 	if err != nil {
 		return nil, err
@@ -33,10 +29,10 @@ func (s *CategoriesService) ListCategories() ([]Category, error) {
 	return toCategoryResponse(categories), nil
 }
 
-func toCategoryResponse(categories []models.Category) []Category {
-	mapped := make([]Category, len(categories))
+func toCategoryResponse(categories []models.Category) []api.CategoryResponse {
+	mapped := make([]api.CategoryResponse, len(categories))
 	for i, c := range categories {
-		mapped[i] = Category{
+		mapped[i] = api.CategoryResponse{
 			Code: c.Code,
 			Name: c.Name,
 		}